Review feedback: remove unused CloudwatchQuery class, just keep quoting logic
